+++ .github/workflows/backend-ci.yml
@@ -2,19 +2,20 @@ name: Backend Continuous Integration
 on:
   push:
     paths:
-      - "backend/**"
+      - "cmd/backend/**"
+      - "pkg/**"
 jobs:
   lint:
     uses: ./.github/workflows/go_lint.yml
     with:
-      working-directory: 'backend'
+      working-directory: 'cmd/backend'
   build:
     needs: lint
     uses: ./.github/workflows/go_build.yml
     with:
-      working-directory: 'backend'
+      working-directory: 'cmd/backend'
   test:
     needs: build
     uses: ./.github/workflows/go_test.yml
     with:
-      working-directory: 'backend'
+      working-directory: 'cmd/backend'
