--- backend/dockerfile
@@ -1,16 +1,0 @@
-FROM arm64v8/golang:1.22.2-alpine as build
-
-WORKDIR /app
-
-COPY . .
-
-RUN GOOS=linux GOARCH=arm64 go build -o backend
-
-RUN apk update && apk add ca-certificates && update-ca-certificates
-
-FROM arm64v8/busybox:1.36.1
-
-COPY --from=build /etc/ssl/certs /etc/ssl/certs
-COPY --from=build /app/backend /app/
-
-CMD [ "/app/backend" ]
