+++ cmd/backend/handlers/create_relationship_and_person.go
@@ -6,20 +6,20 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/vcscsvcscs/GenerationsHeritage/backend/memgraph"
+	"github.com/vcscsvcscs/GenerationsHeritage/pkg/memgraph"
 )
 
-func VerifyRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
+func CreateRelationshipAndPerson(driver neo4j.DriverWithContext) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		var relationship memgraph.Relationship
-		if err := c.ShouldBindJSON(&relationship); err != nil {
+		var rp memgraph.RelationshipAndPerson
+		if err := c.ShouldBindJSON(&rp); err != nil {
 			log.Printf("ip: %s error: %s", c.ClientIP(), err)
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 
 			return
 		}
 
-		rec, err := relationship.VerifyRelationship(driver)
+		rec, err := rp.CreateRelationshipAndPerson(driver)
 		if err != nil {
 			log.Printf("ip: %s error: %s", c.ClientIP(), err)
 			c.JSON(http.StatusInternalServerError, gin.H{"error": "internal server error"})
@@ -27,6 +27,6 @@ func VerifyRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
 			return
 		}
 
-		c.JSON(http.StatusOK, gin.H{"relationship": rec.AsMap()})
+		c.JSON(http.StatusCreated, gin.H{"relationship": rec.AsMap()})
 	}
 }
