+++ cmd/backend/handlers/deletePerson.go
@@ -7,10 +7,10 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/vcscsvcscs/GenerationsHeritage/backend/memgraph"
+	"github.com/vcscsvcscs/GenerationsHeritage/pkg/memgraph"
 )
 
-func CreatePerson(driver neo4j.DriverWithContext) gin.HandlerFunc {
+func DeletePerson(driver neo4j.DriverWithContext) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		if c.Request.Body == nil || c.ContentType() != "application/json" {
 			log.Printf("ip: %s error: request body is empty or content type is not application/json", c.ClientIP())
@@ -22,28 +22,27 @@ func CreatePerson(driver neo4j.DriverWithContext) gin.HandlerFunc {
 		var person memgraph.Person
 		err := json.NewDecoder(c.Request.Body).Decode(&person)
 		if err != nil {
-			log.Printf("ip: %s error: %s", c.ClientIP(), err.Error())
+			log.Printf("ip: %s error: %s", c.ClientIP(), err)
 			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
 
 			return
 		}
 
-		if err := person.Verify(); err != nil {
-			log.Printf("ip: %s error: %s", c.ClientIP(), err.Error())
-			c.JSON(http.StatusBadRequest, gin.H{"error": "contains-forbidden-characters"})
+		if person.ID != "" {
+			log.Printf("ip: %s error: %s", c.ClientIP(), err)
+			c.JSON(http.StatusBadRequest, gin.H{"error": "no person ID provided"})
 
 			return
 		}
 
-		person.ID = c.GetString("id")
-		rec, err := person.CreatePerson(driver)
+		err = person.DeletePerson(driver)
 		if err != nil {
-			log.Printf("ip: %s error: %s", c.ClientIP(), err.Error())
-			c.JSON(http.StatusBadRequest, gin.H{"error": "already-exists"})
+			log.Printf("ip: %s error: %s", c.ClientIP(), err)
+			c.JSON(http.StatusNotFound, gin.H{"error": "could not delete person with ID provided"})
 
 			return
 		}
 
-		c.JSON(http.StatusCreated, gin.H{"person": rec.AsMap()})
+		c.JSON(http.StatusOK, gin.H{"status": "person deleted successfully"})
 	}
 }
