+++ cmd/backend/handlers/updatePerson.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/vcscsvcscs/GenerationsHeritage/backend/memgraph"
+	"github.com/vcscsvcscs/GenerationsHeritage/pkg/memgraph"
 )
 
 func UpdatePerson(driver neo4j.DriverWithContext) gin.HandlerFunc {
