+++ cmd/backend/handlers/verifyRelationship.go
@@ -6,10 +6,10 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/vcscsvcscs/GenerationsHeritage/backend/memgraph"
+	"github.com/vcscsvcscs/GenerationsHeritage/pkg/memgraph"
 )
 
-func DeleteRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
+func VerifyRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		var relationship memgraph.Relationship
 		if err := c.ShouldBindJSON(&relationship); err != nil {
@@ -19,7 +19,7 @@ func DeleteRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
 			return
 		}
 
-		err := relationship.DeleteRelationship(driver)
+		rec, err := relationship.VerifyRelationship(driver)
 		if err != nil {
 			log.Printf("ip: %s error: %s", c.ClientIP(), err)
 			c.JSON(http.StatusInternalServerError, gin.H{"error": "internal server error"})
@@ -27,6 +27,6 @@ func DeleteRelationship(driver neo4j.DriverWithContext) gin.HandlerFunc {
 			return
 		}
 
-		c.JSON(http.StatusAccepted, gin.H{"status": "relationship deleted successfully"})
+		c.JSON(http.StatusOK, gin.H{"relationship": rec.AsMap()})
 	}
 }
