+++ pkg/memgraph/cypher_verify_string.go
