+++ pkg/memgraph/delete_person.go
