+++ pkg/memgraph/model.go
